Remove backbone model and add UserPreferencesModel to components/user folder
